Removed asynchrounous resets.
The design is now more compact.
Removed the ILAs too.
There are no simple ways to remove the critical warning of a LUT being
in the clock tree. The XOR gate will most likely implemented in a LUT.
As a workaround, will have to use a clock wizard IP and use its output
for the ROM. However for now the deployed system works OK.